refactor: 💡 Split styles from `styles.ts` into multiple files
Organized styles files for easier navigation and UI/UX design
refactoring
