+++ screens/calendar/components/CalendarEventsList.tsx
@@ -4,7 +4,7 @@ import { Icon, List, Text } from "react-native-paper";
 import { styles } from "../../../styles/styles";
 import moment from "moment";
 import { ICalendarEvent } from "../../../common/interfaces";
-import { calendarEventsListStyles } from "../../../styles/calendar-events-list-styles";
+import { calendarEventsListStyles } from "../../../styles/calendar-styles/calendar-events-list-styles";
 
 interface ICalendarEventsListProps {
   calendarEvents: ICalendarEvent[];
