+++ styles/app-header-styles.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from "react-native";
+
+export const appHeaderStyles = StyleSheet.create({
+  appHeaderContainer: {
+    marginTop: 50,
+    marginLeft: 15,
+    alignItems: "flex-start",
+    justifyContent: "flex-start",
+  },
+  appHeaderText: {
+    fontSize: 24,
+    fontWeight: "bold",
+    color: "#000000",
+    marginBottom: 5,
+  },
+  appHeaderSubText: {
+    fontSize: 20,
+    fontWeight: "500",
+    color: "#000000",
+  },
+});
