+++ styles/calendar-styles/selected-date-box-styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+
+export const selectedDateBoxStyles = StyleSheet.create({
+  selectedDateText: {
+    alignSelf: "center",
+    fontSize: 16,
+    fontWeight: "bold",
+  },
+});
